Adding disconnect cause to the Android call log.
We'd like to add call disconnect cause codes to the Android call log
content provider.

As part of this work we're also consolidating the definition of call
disconnect codes in a single place. The integer constants defined in
android.telephony.DisconnectCause will be used everywhere and the
following two enums will be removed:
 - android.internal.telephony.Connection.DisconnectCause
 - com.android.services.telephony.common.Call.DisconnectCause

Change-Id: I84a3fd5182a51d07fc7d81076e4b2c15ff26c61c

Conflicts:
	src/com/android/providers/contacts/ContactsDatabaseHelper.java